Conflicts:
	server/inc/UserManager.hh
	server/src/GameManager.cpp

+++ server/inc/IThreadPool.hh
@@ -7,11 +7,11 @@ template <typename RET_VAL, typename ARG>
 class IThreadPool {
 public:
     virtual ~IThreadPool() {}
-    virtual size_t getIdleThreadNumber() const = 0;
-    virtual void joinThread(size_t id) = 0;
+    virtual uint64_t getIdleThreadNumber() const = 0;
+    virtual void joinThread(uint64_t id) = 0;
     virtual void joinAll() = 0;
-    virtual size_t add(RET_VAL(*)(ARG), ARG) = 0;
-    virtual void startThread(size_t id) = 0;
+    virtual uint64_t add(RET_VAL(*)(ARG), ARG) = 0;
+    virtual void startThread(uint64_t id) = 0;
     virtual void startAll() = 0;
 };
